Move IpAddressIdentity internal
